sentinel-guard/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Task queue
celery[redis]==5.4.0
redis==5.1.1

# Scanning
shodan==1.31.0
nmap3==1.6.0
dnspython==2.7.0
httpx==0.27.2
ipwhois==1.3.0

# Crypto & reporting
cryptography==43.0.3

# Rate limiting
slowapi==0.1.9

# Utils
python-dotenv==1.0.1
